+++ src/Classes/FetchedPostData.ts
@@ -1,0 +1,14 @@
+export class FetchedPostData {
+    username!: any
+    caption!: any
+    // post: any
+    post_name!: any
+    post_url!: any
+    keymail!: any
+
+
+    // constructor(post: File) {
+    //     this.post = post
+    // }
+
+}